+ The Last Push
There are only a few TODOs left on the list.  We got a couple done
tonight; new sound effects were added, cutscenes were mostly coded, and
I added some more polish while Sara finishes up the cutscene graphics.
We're going to try to make this week the last push; she's going to focus
on this project, and I'll hop on in the evenings to do what I can to
improve it.  My biggest focus right now is decreasing the load time.  To
taht end I spent most of this evening trying to unwind what part of the
initial load is taking so long.. and it's still not totally clear to me.
But I'll get it.

I'm so excited to finally be finishing a project!