+ Finishg Touches
Sara's been hard at work on the intro/outro cutscenes, and we're getting
really close.  Tonight's update includes two of the three intro panels
finished (the last is almost done) and some additional graphical tweaks,
namely the kitchen and bedroom got a bit more door on the right side to
compensate for screen space hidden by system bars on android devices.

Once the last cutscene panels are done, we're going to review the entire
game's dialog to ensure it's consistent and call this done!