+ Final In-Game Graphics Complete
The title says it all!  We've now finished all the in-game graphic!  The
only remaining things are the cutscene cards; one for the intro needs
color, and the outro needs to be finished and added to the game.

Tonight we added the sleeping sprites for Frank.  We also added Red, a
new character who's going to handle reminder dialogs; soon, instead of
talking to yourself when leaving, you'll be talking to Red.  He'll also
give hints, so if you get stuck he can rmind you what you still need to
do, and also give you some hints about how to do it.

But the real important question: are we any closer to coming up with a
name for this project?

No, no we're not.