+ Meet Red!
I'd thought that all the programming was done, but a last-minute
addition required a bit more - introducing red, your nanny pat friend
that is also the game's hint system.

We were already talking about putting him in - he's part of the book
Sara's writing that this is based off of, and his omission would be
weird.  In the book, red is the one keeping an eye on you while Dad's
out, and so we decided that he would make a good hint system in the
game.

The hint system for this game has been in flux for a while.  For a few
builds, a "?" icon appeared in the bottom right corner, and highlighted
all interactive elements in a room when you hovered the mouse over it.
That wasn't awesome, though - it often drew your attention to irrelevant
things, and the game doesn't have any real pixel-hunting anyway.  It
didn't solve the problem of "I'm stuck and don't know what to do," it
just hightlighted things to click on and hoped the player would
eventually click the right things.

Red now provides actual hints based on where you are in the game.  Since
you can have multiple quests going at once, I structed these as you
asking red for advince on something, or maybe just complaining about a
setback, with red responding with something helpful.  Which quest is
chosen for any individual interaction is random, maybe helping players
stop persuing a quest they're stuck on to try something new that they
might have some success with (and hopefully figuring out something in
their current quest while they're at it).

I'm pretty sure that, past some minor work to get cutscenes fully
finished, this is the last of the code.