+ Pluck
Tonight, Sara finished up the cutscene while I coded them, and at long
last all the assets and code are finished.  We've even got a thumbnail:

![thumbnail](https://raw.githubusercontent.com/Dorthu/2019-christmas-adventure/master/meta/thumbnail.png)

Even more exciting, we have a name:

_Pluck_

Sara came up with it, which means she did the hardest part of this
whole project.  The word "pluck" is commonly used to describe taking
something quickly, but it has another definition - bravery or courage.

We're going to review all of the text in the game soon, and then send it
out to a round of tester before calling it finished.  I'm very excited.