+ Finishing Touches
With the project finished, there was still a bit more to actually _do_.

There always seems to be.

So, we cleaned up the [itch.io page](https://dorthu.itch.io/pluck), and
I added all the metadata I could to the project here for safekeeping.
At this point, all I need to do is get the word out there that this is a
thing that exists, and people should check it out.

..where to go from here?

I don't know.  I never thought I'd make it this far.  It's a finished
project.  A whole new adventure.